display: dsi-panel-n16: Correct physical panel dimensions
These values were multiplied with 10 by xiaomi to allow
more specific configurations. This requires changes in
the display HAL which are not part of the open source
qcom display HAL. Hence decrease the values by dividing
through 10 and rounding to nearest integer to support
the qcom display HAL.

This caused unexpected behaviour in some apps without the
display HAL changes. (e.g. pictures are shown too small
in Reddit app, sliding in Gboard is too sensitive)

It is assumed that the value is too high if the height
is above 1000. The changes have been done with this script:
```
for panel in $(ls qcom/display/display/dsi-panel-{l,m}*); do
        height=$(sed -n "s/^.*qcom,mdss-pan-physical-height-dimension = <\([0-9]\+\)>;/\1/p" ${panel})
        width=$(sed -n "s/^.*qcom,mdss-pan-physical-width-dimension = <\([0-9]\+\)>;/\1/p" ${panel})

        if [[ ${height} > 1000 ]]; then
                height=$(python3 -c "print(round(${height} / 10))")
                width=$(python3 -c "print(round(${width} / 10))")

                sed -i "s/^\(.*qcom,mdss-pan-physical-height-dimension = <\)[0-9]\+\(>;\)/\1${height}\2/g" ${panel}
                sed -i "s/^\(.*qcom,mdss-pan-physical-width-dimension = <\)[0-9]\+\(>;\)/\1${width}\2/g" ${panel}
        fi
done
```

Change-Id: Ia99595bf038d8f9302e36e5a02aae48912243c96